Used React.ComponentPropsWithoutRef
