+++ src/components/dropdown/EllipsisDropdown.tsx
@@ -1,35 +1,12 @@
-import React from 'react';
 import { Dropdown } from 'react-bootstrap';
-import { DropdownToggleProps } from 'react-bootstrap/esm/DropdownToggle';
 
-import { faEllipsisV } from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import CustomToggle from './Toggle';
 
 type Props = {
   header: string;
   items: { label: string; value: string }[];
   onClick: (code: string) => void;
 };
-
-type CustomToggleProps = DropdownToggleProps & {
-  onClick: () => void;
-};
-
-const CustomToggle = React.forwardRef<HTMLButtonElement, CustomToggleProps>(
-  ({ onClick }: CustomToggleProps, ref) => (
-    <button
-      className="border-0 bg-transparent"
-      ref={ref}
-      onClick={(e) => {
-        e.preventDefault();
-        onClick(e);
-      }}
-    >
-      <FontAwesomeIcon icon={faEllipsisV} />
-    </button>
-  )
-);
-CustomToggle.displayName = 'CustomToggle';
 
 const EllipsisDropdown = ({ header, items, onClick }: Props) => (
   <Dropdown>
