+++ src/components/tile/OptionsTile.tsx
@@ -5,30 +5,17 @@ import { EllipsisDropdown } from '@/components';
 
 import './Tile.scss';
 
-type Props = {
+type Props = React.ComponentPropsWithoutRef<typeof EllipsisDropdown> & {
   tileHeader: string;
   children: React.ReactNode;
-  dropdownHeader: string;
-  dropdownItems: { label: string; value: string }[];
-  dropdownOnClick: (code: string) => void;
 };
 
-const OptionsTile = ({
-  tileHeader,
-  children,
-  dropdownHeader,
-  dropdownItems,
-  dropdownOnClick,
-}: Props) => {
+const OptionsTile = ({ tileHeader, children, ...dropdownProps }: Props) => {
   return (
     <Card>
       <Card.Header className="font-weight-bold text-primary d-flex justify-content-between">
         {tileHeader}
-        <EllipsisDropdown
-          header={dropdownHeader}
-          items={dropdownItems}
-          onClick={dropdownOnClick}
-        />
+        <EllipsisDropdown {...dropdownProps} />
       </Card.Header>
       <Card.Body className="d-flex flex-column gap-4">{children}</Card.Body>
     </Card>
